-Complétion du code sur l'itérateur : ajout de filtrage sur les indices des lignes et des colonnes

+++ trunk/AMAPVox/Commons/src/main/java/fr/amap/amapvox/commons/util/SphericalCoordinates.java
@@ -35,7 +35,7 @@ public class SphericalCoordinates {
     public SphericalCoordinates() {
     }
 
-    public SphericalCoordinates(float azimuth, float elevation) {
+    public SphericalCoordinates(double azimuth, double elevation) {
         this.azimuth = azimuth;
         this.elevation = elevation;
     }
